requirements.txt
python-dotenv

llama-index
openai
pydantic


